streams: encode time gaps that do not fit in an int16
A query died mid-response with "Python integer 32967 out of bounds for
int16", out of rld_encode. The stream was fine: it was an activity left
recording for nine hours after it ended, so one diff in its time stream
was 32967 seconds, and the codec's widest type was int16. Added type 3,
int32 diffs with a 32-bit first value, and the width is now chosen so
that the first value fits too -- the old test only looked at the diffs,
so a first value over 32767 would have raised the same OverflowError
from the header it never wrote.

Time is 32-bit the rest of the way down as well. A Uint16Array holds
18.2 hours and the activity that raised this is already at 33k seconds,
so anything longer was wrapping silently rather than failing. That costs
two bytes a point in the packed stream buffer.

pack() drops an activity it cannot encode instead of raising. That is
what turned one bad stream into the second error in the log -- "the
error response will not be sent to the client... a previous response has
at least partially been sent" -- a truncated body and no way to say so.

Streams already in Mongo are types 0-2 and decode unchanged, but the
frontend has to go out with this: an old bundle would read a type-3 blob
as a 3-byte header and garbage diffs.

Verified the production shape end to end through encode_streams, and
cross-checked the TypeScript decoder against Python-produced bytes for
all four types and ten random streams, at zero and non-zero byteOffset.
New codec tests round-trip each type, the 32967-second gap, the big
first value and the -32768 marker collision.

+++ frontend/src/js/DataImport.ts
@@ -75,7 +75,9 @@ export type ImportedActivity = {
   mpk?: Uint8Array
 
   streams?: {
-    time: Uint16Array
+    /** seconds since the activity started. 32-bit: an activity left recording
+     * overnight runs past the 18 hours a Uint16 can hold. */
+    time: Uint32Array
     altitude: Int16Array
     latlng: Float32Array
   }
@@ -167,7 +169,7 @@ export function decodePackedStreams(
 ): ImportedActivity["streams"] {
   const mpk = <UnpackedStreams>decode(packed)
   return {
-    time: rld_decode(mpk.t, Uint16Array),
+    time: rld_decode(mpk.t, Uint32Array),
     altitude: rld_decode(mpk.a, Int16Array),
     latlng: decode2Buf(mpk.p, polylinePrecision),
   }
